Add REMOVE_TODO for immutable object
Need to develop method to identify presence of target item, and, if
present, get its index and return a new immutable array with the item
'sliced' from it. Combined arr #some and #findIndex to accomplish
the task at hand.  Will need to refactor once a cleaner solution is
found for the task.  Fully test supported with deep-freeze