Set up test for store elements through Todo component
The Provider is wrapped around the <App> component but allows store
to connect through the Todo element (I believe).  Still need to
evaluate the best method for exporting a collection of Actions.
Now, passing each method in individually vs " import * as xyz from"
The latter requires each fn to be called as xyz.fn, which is not as
self documenting as including the functions in the import