Justifying the todo list items
Presenting the results in a <ul> without <li>'s in the map fn
resulted in a leftside tab for the todo items, difficult to justify
Removing the ul element fixed the issue.  Was concerned about
delivering li's without the ul nesting but realized TodoItems
are generated as boostrap/react elements instead.