Create localStorage to persist data & refactor store
Utilize Abaramov's stategy for local storage to persist data. The
store is in a separate file however so config a bit differently.
Needed to create store first in order to enable the #saveState method

LocalStorage now saves the state of the todo app upon refresh.  Need
to continue to update this application for othe contingencies (incl.
the use of additional reducers that should NOT save state).

Here the URL for reference:https://egghead.io/lessons/
javascript-redux-persisting-the-state-to-the-local-storage