Complete self item todo item without need for prompt
Inserted the task as a form item with a default value of item.task.
DefaultValue allows the input to prepopulate with a mutable string
value.  Wrapped in a Form to allow onSubmit event.  Event takes
the current task via refs and checks for a change.  If none, no
action passed to the reducer, otherwise, runs through the Update
props

Allows elimiantaion of the Edit button and now each task element is
self referencing it's poriton of component without needed to 'route'
to another screen or SPA a new form item.  Need to consdider other
UI indicators of 'edit' mode (e.g. color change while in task)