npm ERR! Test failed.  See above for more details.
Setting up get /todos/:id test to pass.  Need to ensure
that the done() method is INSIDE of the .end() in the 'it'
clause, otherwise mocha will terminate on test failure,
meaing the -w watch flag is busted.

.send() will be an important inclusion in the suite if you
are following the attached tutorial.  Removed in this
commit as we walk through a debug.

Remaining test suites may not be working.  Will resolve
after setting this test

Postman shows that the route is working.  Trying to get the
test to support that

In express, it's important that the route show ':id' and not
a string in it's place.  Otherwise it won't recognize it as
a paramater in the http header. This was formerly a broken
issue/bug in previous commits.